move files into a package
