+++ appengine-cron-emulator/src/main/java/org/broad/Cron.java
@@ -1,5 +1,7 @@
+package org.broad;
+
 /*
- * POJOs to represent the Cron config as defined in
+ * POJOs to represent the org.broad.Cron config as defined in
  * https://cloud.google.com/appengine/docs/standard/java/config/cronref#example
  *
  * This class will be fed into jackson xml mapper to parse `cron.xml` files
