+++ appengine-cron-emulator/src/main/java/org/broad/CronEntries.java
@@ -1,3 +1,5 @@
+package org.broad;
+
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlElementWrapper;
 
 import java.util.ArrayList;
